# Optimizing the PlayerSelectSpawn function