--- a/garrysmod/gamemodes/base/gamemode/init.lua
+++ b/garrysmod/gamemodes/base/gamemode/init.lua
@@ -19,6 +19,16 @@ function GM:InitPostEntity()
 end
 
 --[[---------------------------------------------------------
+	Name: gamemode:PostCleanupMap()
+	Desc: Called as after the map has been cleaned up
+-----------------------------------------------------------]]
+function GM:PostCleanupMap()
+
+	self:LoadSpawnPoints()
+
+end
+
+--[[---------------------------------------------------------
 	Name: gamemode:Think()
 	Desc: Called every frame
 -----------------------------------------------------------]]