@callumok2004 2025-06-27
> What about doing 1 loop over ents.GetAll() or ents.Iterator and check the classname of each entity, instead of doing 20 FindByClass calls?

https://github.com/Facepunch/garrysmod/pull/2295 solves this