@Vitroze 2025-06-27
> What about doing 1 loop over ents.GetAll() or ents.Iterator and check the classname of each entity, instead of doing 20 FindByClass calls?

That's right, I'm going to change that - thanks for the comment!